+++ src/dfa/codegen.rs
@@ -191,6 +191,9 @@ pub fn reify(
             // Current lexer state
             state: usize,
 
+            // Set after end-of-input is handled by a rule, or by default in `Init` rule
+            done: bool,
+
             // Which lexer state to switch to on successful match
             initial_state: usize,
 
@@ -261,6 +264,7 @@ pub fn reify(
             #visibility fn new_with_state(input: &'input str, user_state: #user_state_type) -> Self {
                 #lexer_name {
                     state: 0,
+                    done: false,
                     initial_state: 0,
                     user_state: Default::default(),
                     input,
@@ -283,6 +287,10 @@ pub fn reify(
             type Item = Result<(usize, #token_type, usize), LexerError<#(#user_error_type_lifetimes),*>>;
 
             fn next(&mut self) -> Option<Self::Item> {
+                if self.done {
+                    return None;
+                }
+
                 loop {
                     // println!("state = {:?}, next char = {:?}", self.state, self.iter.peek());
                     match self.state {
@@ -356,7 +364,8 @@ fn generate_state_arm(
         initial,
         char_transitions,
         range_transitions,
-        fail_transition,
+        any_transition,
+        end_of_input_transition,
         accepting,
         predecessors: _,
     } = state;
@@ -399,23 +408,44 @@ fn generate_state_arm(
         // stream ends. In non-initial states EOS (end-of-stream) returns the last (longest)
         // match, or fails (error).
 
-        // Backtrack or fail with an error if we don't have a fail transition
-        let fail_action = fail_transition
+        // When we can't take char or range transitions, take the 'any' transition if it exists, or
+        // fail (backtrack or raise error)
+        let default_action = any_transition
             .as_ref()
-            .map(|fail_transition| generate_fail_transition(ctx, states, true, fail_transition))
+            .map(|any_transition| generate_any_transition(ctx, states, true, any_transition))
             .unwrap_or_else(|| fail(ctx));
+
+        let end_of_input_action = match end_of_input_transition {
+            Some(end_of_input_transition) => match end_of_input_transition {
+                Trans::Accept(action) => {
+                    let action = generate_rhs_code(ctx, *action);
+                    quote!(#action)
+                }
+                Trans::Trans(next_state) => {
+                    let StateIdx(next_state) = ctx.renumber_state(*next_state);
+                    quote!(self.state = #next_state;)
+                }
+            },
+            None => {
+                // End-of-input is handled in state 0 by default
+                quote!(return None;)
+            }
+        };
 
         let state_char_arms = generate_state_char_arms(
             ctx,
             states,
             char_transitions,
             range_transitions,
-            &fail_action,
+            &default_action,
         );
 
         quote!(
             match self.iter.peek().copied() {
-                None => return None,
+                None => {
+                    self.done = true; // don't handle end-of-input again
+                    #end_of_input_action
+                }
                 Some((char_idx, char)) => {
                     let char_idx = self.iter_byte_idx + char_idx;
                     self.current_match_start = char_idx;
@@ -428,40 +458,65 @@ fn generate_state_arm(
         )
     } else if let Some(rhs) = accepting {
         // Accepting state
-        let action = generate_rhs_code(ctx, *rhs);
+        let default_action = any_transition
+            .as_ref()
+            .map(|any_transition| generate_any_transition(ctx, states, true, any_transition))
+            .unwrap_or_else(|| generate_rhs_code(ctx, *rhs));
 
-        if char_transitions.is_empty() && range_transitions.is_empty() {
-            quote!({
-                #action
-            })
-        } else {
-            let state_char_arms =
-                generate_state_char_arms(ctx, states, char_transitions, range_transitions, &action);
-
-            let semantic_fn = rhs.symbol();
-
-            // TODO: Braces can be removed in some cases
-            quote!({
-                self.last_match =
-                    Some((self.current_match_start, #semantic_fn, self.current_match_end));
-
-                match self.iter.peek().copied() {
-                    None => {
+        let end_of_input_action = match end_of_input_transition {
+            Some(end_of_input_transition) => match end_of_input_transition {
+                Trans::Accept(action) => {
+                    let action = generate_rhs_code(ctx, *action);
+                    quote!(
+                        self.done = true; // don't handle end-of-input again
                         #action
-                    }
-                    Some((char_idx, char)) => {
-                        match char {
-                            #(#state_char_arms,)*
-                        }
+                    )
+                }
+                Trans::Trans(next_state) => {
+                    let StateIdx(next_state) = ctx.renumber_state(*next_state);
+                    quote!(
+                        self.done = true; // don't handle end-of-input again
+                        self.state = #next_state;
+                    )
+                }
+            },
+            None => {
+                // No `self.done = true` here, end-of-input is not handled
+                generate_rhs_code(ctx, *rhs)
+            }
+        };
+
+        let state_char_arms = generate_state_char_arms(
+            ctx,
+            states,
+            char_transitions,
+            range_transitions,
+            &default_action,
+        );
+
+        let semantic_fn = rhs.symbol();
+
+        // TODO: Braces can be removed in some cases
+        quote!({
+            self.last_match =
+                Some((self.current_match_start, #semantic_fn, self.current_match_end));
+
+            match self.iter.peek().copied() {
+                None => {
+                    #end_of_input_action
+                }
+                Some((char_idx, char)) => {
+                    match char {
+                        #(#state_char_arms,)*
                     }
                 }
-            })
-        }
+            }
+        })
     } else {
         // Non-accepting state
-        let fail_action = fail_transition
+        let default_action = any_transition
             .as_ref()
-            .map(|fail_transition| generate_fail_transition(ctx, states, *initial, fail_transition))
+            .map(|any_transition| generate_any_transition(ctx, states, *initial, any_transition))
             .unwrap_or_else(|| fail(ctx));
 
         let state_char_arms = generate_state_char_arms(
@@ -469,22 +524,42 @@ fn generate_state_arm(
             states,
             char_transitions,
             range_transitions,
-            &fail_action,
+            &default_action,
         );
 
-        let end_of_stream_action = if *initial {
-            // In an initial state other than the state 0 we fail with "unexpected EOF"
-            let error = make_lexer_error();
-            quote!(return Some(Err(#error));)
-        } else {
-            // Otherwise we run the fail action and go to initial state of the current DFA. Initial
-            // state will then fail.
-            fail_action
+        let end_of_input_action = match end_of_input_transition {
+            Some(end_of_input_transition) => match end_of_input_transition {
+                Trans::Accept(action) => {
+                    let action = generate_rhs_code(ctx, *action);
+                    quote!(
+                        self.done = true; // don't handle end-of-input again
+                        #action
+                    )
+                }
+                Trans::Trans(next_state) => {
+                    let StateIdx(next_state) = ctx.renumber_state(*next_state);
+                    quote!(
+                        self.done = true; // don't handle end-of-input again
+                        self.state = #next_state;
+                    )
+                }
+            },
+            None => {
+                if *initial {
+                    // In an initial state other than state 0 we fail with "unexpected EOF" error
+                    let error = make_lexer_error();
+                    quote!(return Some(Err(#error));)
+                } else {
+                    // Otherwise we run the fail action and go to initial state of the current DFA.
+                    // Initial state will then fail.
+                    default_action.clone()
+                }
+            }
         };
 
         quote!(match self.iter.peek().copied() {
             None => {
-                #end_of_stream_action
+                #end_of_input_action
             }
             Some((_, char)) => {
                 match char {
@@ -495,7 +570,7 @@ fn generate_state_arm(
     }
 }
 
-fn generate_fail_transition(
+fn generate_any_transition(
     ctx: &mut CgCtx,
     states: &[State<Trans, SemanticActionIdx>],
     initial: bool,
@@ -508,6 +583,8 @@ fn generate_fail_transition(
             } else {
                 let StateIdx(next_state) = ctx.renumber_state(StateIdx(*next_state));
                 quote!(
+                    self.current_match_end += char.len_utf8();
+                    let _ = self.iter.next();
                     self.state = #next_state;
                 )
             }
@@ -534,7 +611,8 @@ fn generate_state_char_arms(
     states: &[State<Trans, SemanticActionIdx>],
     char_transitions: &FxHashMap<char, Trans>,
     range_transitions: &RangeMap<Trans>,
-    fail_action: &TokenStream,
+    // RHS of the default alternative for this `match` (_ => <default_rhs>)
+    default_rhs: &TokenStream,
 ) -> Vec<TokenStream> {
     // Arms of the `match` for the current character
     let mut state_char_arms: Vec<TokenStream> = vec![];
@@ -635,7 +713,7 @@ fn generate_state_char_arms(
         ));
     }
 
-    state_char_arms.push(quote!(_ => { #fail_action }));
+    state_char_arms.push(quote!(_ => { #default_rhs }));
 
     state_char_arms
 }